Move remaining peripheral tests to src
